+++ application/Espo/Jobs/Cleanup.php
@@ -33,23 +33,27 @@ use \Espo\Core\Exceptions;
 
 class Cleanup extends \Espo\Core\Jobs\Base
 {
-    protected $cleanupJobPeriod = '1 month';
+    protected $cleanupJobPeriod = '10 days';
 
     protected $cleanupActionHistoryPeriod = '15 days';
 
     protected $cleanupAuthTokenPeriod = '1 month';
 
+    protected $cleanupAuthLogPeriod = '2 months';
+
     protected $cleanupNotificationsPeriod = '2 months';
 
     protected $cleanupRemovedNotesPeriod = '2 months';
 
-    protected $cleanupAttachmentsPeriod = '1 month';
+    protected $cleanupAttachmentsPeriod = '15 days';
 
     protected $cleanupAttachmentsFromPeriod = '3 months';
 
     protected $cleanupRemindersPeriod = '15 days';
 
     protected $cleanupBackupPeriod = '2 month';
+
+    protected $cleanupDeletedRecordsPeriod = '3 months';
 
     public function run()
     {
@@ -61,8 +65,10 @@ class Cleanup extends \Espo\Core\Jobs\Base
         $this->cleanupNotifications();
         $this->cleanupActionHistory();
         $this->cleanupAuthToken();
+        $this->cleanupAuthLog();
         $this->cleanupUpgradeBackups();
         $this->cleanupUniqueIds();
+        $this->cleanupDeletedRecords();
     }
 
     protected function cleanupJobs()
@@ -70,7 +76,10 @@ class Cleanup extends \Espo\Core\Jobs\Base
         $pdo = $this->getEntityManager()->getPDO();
 
         $query = "DELETE FROM `job` WHERE DATE(modified_at) < ".$pdo->quote($this->getCleanupJobFromDate())." AND status <> 'Pending'";
+        $sth = $pdo->prepare($query);
+        $sth->execute();
 
+        $query = "DELETE FROM `job` WHERE DATE(modified_at) < ".$pdo->quote($this->getCleanupJobFromDate())." AND status = 'Pending' AND deleted = 1";
         $sth = $pdo->prepare($query);
         $sth->execute();
     }
@@ -155,6 +164,20 @@ class Cleanup extends \Espo\Core\Jobs\Base
         $sth->execute();
     }
 
+    protected function cleanupAuthLog()
+    {
+        $pdo = $this->getEntityManager()->getPDO();
+
+        $period = '-' . $this->getConfig()->get('cleanupAuthLogPeriod', $this->cleanupAuthLogPeriod);
+        $datetime = new \DateTime();
+        $datetime->modify($period);
+
+        $query = "DELETE FROM `auth_log_record` WHERE DATE(created_at) < " . $pdo->quote($datetime->format('Y-m-d')) . "";
+
+        $sth = $pdo->prepare($query);
+        $sth->execute();
+    }
+
     protected function getCleanupJobFromDate()
     {
         $period = '-' . $this->getConfig()->get('cleanupJobPeriod', $this->cleanupJobPeriod);
@@ -174,11 +197,11 @@ class Cleanup extends \Espo\Core\Jobs\Base
         $collection = $this->getEntityManager()->getRepository('Attachment')->where(array(
             'OR' => array(
                 array(
-                    'role' => ['Export File']
+                    'role' => ['Export File', 'Mail Merge', 'Mass Pdf']
                 )
             ),
             'createdAt<' => $datetime->format('Y-m-d H:i:s')
-        ))->limit(0, 1000)->find();
+        ))->limit(0, 5000)->find();
 
         foreach ($collection as $e) {
             $this->getEntityManager()->removeEntity($e);
@@ -203,7 +226,7 @@ class Cleanup extends \Espo\Core\Jobs\Base
                 ),
                 'createdAt<' => $datetime->format('Y-m-d H:i:s'),
                 'createdAt>' => '2017-05-10 00:00:00'
-            ))->limit(0, 1000)->find();
+            ))->limit(0, 5000)->find();
 
             foreach ($collection as $e) {
                 $this->getEntityManager()->removeEntity($e);
@@ -235,25 +258,30 @@ class Cleanup extends \Espo\Core\Jobs\Base
             }
             if (!$hasAttachmentField) continue;
 
-            $deletedEntityList = $this->getEntityManager()->getRepository($scope)->where([
+            if (!$this->getEntityManager()->hasRepository($scope)) continue;
+            $repository = $this->getEntityManager()->getRepository($scope);
+            if (!method_exists($repository, 'find')) continue;
+            if (!method_exists($repository, 'where')) continue;
+
+            $deletedEntityList = $repository->where([
                 'deleted' => 1,
                 'modifiedAt<' => $datetime->format('Y-m-d H:i:s'),
                 'modifiedAt>' => $datetimeFrom->format('Y-m-d H:i:s'),
 
             ])->find(['withDeleted' => true]);
             foreach ($deletedEntityList as $deletedEntity) {
-                $attachmentToRemoveList = $this->getEntityManager()->getRepository('Attachment')->where(array(
-                    'OR' => array(
-                        array(
+                $attachmentToRemoveList = $this->getEntityManager()->getRepository('Attachment')->where([
+                    'OR' => [
+                        [
                             'relatedType' => $scope,
                             'relatedId' => $deletedEntity->id
-                        ),
-                        array(
+                        ],
+                        [
                             'parentType' => $scope,
                             'parentId' => $deletedEntity->id
-                        )
-                    )
-                ))->find();
+                        ]
+                    ]
+                ])->find();
 
                 foreach ($attachmentToRemoveList as $attachmentToRemove) {
                     $this->getEntityManager()->removeEntity($attachmentToRemove);
@@ -348,6 +376,69 @@ class Cleanup extends \Espo\Core\Jobs\Base
                 $info = new \SplFileInfo($dirPath);
                 if ($datetime->getTimestamp() > $info->getMTime()) {
                     $fileManager->removeInDir($dirPath, true);
+                }
+            }
+        }
+    }
+
+    protected function cleanupDeletedRecords()
+    {
+        if (!$this->getConfig()->get('cleanupDeletedRecords')) return;
+        $period = '-' . $this->getConfig()->get('cleanupDeletedRecordsPeriod', $this->cleanupDeletedRecordsPeriod);
+        $datetime = new \DateTime('-' . $period);
+
+        $query = $this->getEntityManager()->getQuery();
+
+        $scopeList = array_keys($this->getMetadata()->get(['scopes']));
+        foreach ($scopeList as $scope) {
+            if (!$this->getMetadata()->get(['scopes', $scope, 'entity'])) continue;
+            if ($scope === 'Attachment') continue;
+            if (!$this->getMetadata()->get(['entityDefs', $scope, 'fields', 'modifiedAt'])) continue;
+
+            if (!$this->getEntityManager()->hasRepository($scope)) continue;
+            $repository = $this->getEntityManager()->getRepository($scope);
+            if (!$repository) continue;
+            if (!method_exists($repository, 'find')) continue;
+            if (!method_exists($repository, 'where')) continue;
+            if (!method_exists($repository, 'select')) continue;
+            if (!method_exists($repository, 'deleteFromDb')) continue;
+
+            $deletedEntityList = $repository->select(['id', 'deleted'])->where([
+                'deleted' => 1,
+                'modifiedAt<' => $datetime->format('Y-m-d H:i:s')
+            ])->find(['withDeleted' => true]);
+            foreach ($deletedEntityList as $e) {
+                if (!$e->get('deleted')) continue;
+                $repository->deleteFromDb($e->id);
+
+                foreach ($e->getRelationList() as $relation) {
+                    if ($e->getRelationType($relation) !== 'manyMany') continue;
+                    try {
+                        $relationName = $e->getRelationParam($relation, 'relationName');
+                        $relationTable = $query->toDb($relationName);
+
+                        $midKey = $e->getRelationParam($relation, 'midKeys')[0];
+
+                        $where = [];
+                        $where[$midKey] = $e->id;
+
+                        $conditions = $e->getRelationParam($relation, 'conditions');
+                        if (!empty($conditions)) {
+                            foreach ($conditions as $key => $value) {
+                                $where[$key] = $value;
+                            }
+                        }
+
+                        $partList = [];
+                        foreach ($where as $key => $value) {
+                            $partList[] = $query->toDb($key) . ' = ' . $query->quote($value);
+                        }
+                        if (empty($partList)) continue;
+
+                        $sql = "DELETE FROM `{$relationTable}` WHERE " . implode(' AND ', $partList);
+
+                        $this->getEntityManager()->getPDO()->query($sql);
+                    } catch (\Exception $e) {}
                 }
             }
         }
